paddle-ocr-sidecar/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
paddlepaddle==3.0.0
paddleocr==2.9.1
Pillow==12.1.1
numpy==1.26.4
python-multipart==0.0.22
pypdfium2==4.30.1
pillow-heif==0.18.0
